requirements.txt
flask>=3.0.0
psycopg2-binary>=2.9.9
pandas>=2.1.4
numpy>=1.26.2
scikit-learn>=1.3.2
xgboost>=2.0.3
joblib>=1.3.2
pytz>=2024.1
python-dotenv>=1.0.0
werkzeug>=3.0.0
